initial cut at timestamp_wonkiness handling for Lotek clock jumps
This attempt isn't complete:  it doesn't add the additional paths
in the DFA graph; it just implements the filtering for time drift.

This approach seems clunky; it separates the accept-pulse-with-a-clock-jump
logic from the don't drift too far logic in an awkward way.

(We'll try another that just modifies the DFA graph instead.)

- clean up logic of pulse acceptance:  a candidate at any tag_id_level can
accept a pulse without claiming ownership.  When ownership is claimed,
any other candidates are barred from accepting that pulse, and any which
have already (but not claimed ownership) are deleted.

- check for time drift between start of run and current burst; this is
the logic that prevents